Escape <!--, --> and <script (#20)
In HTML script elements, `</script` is not the only substring that
switches the HTML parser state even when contained in a JS string
literal. This commit adds escaping for `<!--`, `-->` and `<script`,
which appear to be all substrings that cause state transitions which
are not reset by the `"` closing the string literal.